requirementsSTT.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==2.11.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-speech==2.16.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.1
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
keyboard==0.13.5
numpy==1.24.1
oauthlib==3.2.2
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pydub==0.25.1
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
websockets==10.4
